+++ csas-20230120/alibabacloud_csas20230120/models/_get_user_device_response_body.py
@@ -50,7 +50,12 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         app_status: str = None,
         app_version: str = None,
         auto_login_status: str = None,
+        battery_health_percentage: int = None,
+        battery_remaining_percentage: int = None,
         cpu: str = None,
+        city: str = None,
+        continent: str = None,
+        country: str = None,
         create_time: str = None,
         department: str = None,
         device_belong: str = None,
@@ -60,18 +65,23 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         device_type: str = None,
         device_version: str = None,
         disk: str = None,
+        disk_available: int = None,
+        disk_used: int = None,
         dlp_status: str = None,
         edr_status: str = None,
         history_users: List[main_models.GetUserDeviceResponseBodyDeviceHistoryUsers] = None,
         hostname: str = None,
         ia_status: str = None,
         inner_ip: str = None,
+        join_ad_domain: bool = None,
         mac: str = None,
         match_device_group_ids: List[str] = None,
         memory: str = None,
         nac_status: str = None,
         net_interface_info: List[main_models.GetUserDeviceResponseBodyDeviceNetInterfaceInfo] = None,
         pa_status: str = None,
+        processes: List[main_models.GetUserDeviceResponseBodyDeviceProcesses] = None,
+        province: str = None,
         sase_user_id: str = None,
         sharing_status: bool = None,
         sn_base_board: str = None,
@@ -80,6 +90,7 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         sn_processor: str = None,
         sn_system: str = None,
         src_ip: str = None,
+        terminal_info_collect_time: int = None,
         update_time: str = None,
         username: str = None,
         workshop: str = None,
@@ -87,7 +98,12 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         self.app_status = app_status
         self.app_version = app_version
         self.auto_login_status = auto_login_status
+        self.battery_health_percentage = battery_health_percentage
+        self.battery_remaining_percentage = battery_remaining_percentage
         self.cpu = cpu
+        self.city = city
+        self.continent = continent
+        self.country = country
         self.create_time = create_time
         self.department = department
         self.device_belong = device_belong
@@ -97,18 +113,23 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         self.device_type = device_type
         self.device_version = device_version
         self.disk = disk
+        self.disk_available = disk_available
+        self.disk_used = disk_used
         self.dlp_status = dlp_status
         self.edr_status = edr_status
         self.history_users = history_users
         self.hostname = hostname
         self.ia_status = ia_status
         self.inner_ip = inner_ip
+        self.join_ad_domain = join_ad_domain
         self.mac = mac
         self.match_device_group_ids = match_device_group_ids
         self.memory = memory
         self.nac_status = nac_status
         self.net_interface_info = net_interface_info
         self.pa_status = pa_status
+        self.processes = processes
+        self.province = province
         self.sase_user_id = sase_user_id
         self.sharing_status = sharing_status
         self.sn_base_board = sn_base_board
@@ -117,6 +138,7 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         self.sn_processor = sn_processor
         self.sn_system = sn_system
         self.src_ip = src_ip
+        self.terminal_info_collect_time = terminal_info_collect_time
         self.update_time = update_time
         self.username = username
         self.workshop = workshop
@@ -128,6 +150,10 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
                     v1.validate()
         if self.net_interface_info:
             for v1 in self.net_interface_info:
+                 if v1:
+                    v1.validate()
+        if self.processes:
+            for v1 in self.processes:
                  if v1:
                     v1.validate()
 
@@ -145,8 +171,23 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         if self.auto_login_status is not None:
             result['AutoLoginStatus'] = self.auto_login_status
 
+        if self.battery_health_percentage is not None:
+            result['BatteryHealthPercentage'] = self.battery_health_percentage
+
+        if self.battery_remaining_percentage is not None:
+            result['BatteryRemainingPercentage'] = self.battery_remaining_percentage
+
         if self.cpu is not None:
             result['CPU'] = self.cpu
+
+        if self.city is not None:
+            result['City'] = self.city
+
+        if self.continent is not None:
+            result['Continent'] = self.continent
+
+        if self.country is not None:
+            result['Country'] = self.country
 
         if self.create_time is not None:
             result['CreateTime'] = self.create_time
@@ -175,6 +216,12 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         if self.disk is not None:
             result['Disk'] = self.disk
 
+        if self.disk_available is not None:
+            result['DiskAvailable'] = self.disk_available
+
+        if self.disk_used is not None:
+            result['DiskUsed'] = self.disk_used
+
         if self.dlp_status is not None:
             result['DlpStatus'] = self.dlp_status
 
@@ -195,6 +242,9 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         if self.inner_ip is not None:
             result['InnerIP'] = self.inner_ip
 
+        if self.join_ad_domain is not None:
+            result['JoinAdDomain'] = self.join_ad_domain
+
         if self.mac is not None:
             result['Mac'] = self.mac
 
@@ -214,6 +264,14 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
 
         if self.pa_status is not None:
             result['PaStatus'] = self.pa_status
+
+        result['Processes'] = []
+        if self.processes is not None:
+            for k1 in self.processes:
+                result['Processes'].append(k1.to_map() if k1 else None)
+
+        if self.province is not None:
+            result['Province'] = self.province
 
         if self.sase_user_id is not None:
             result['SaseUserId'] = self.sase_user_id
@@ -239,6 +297,9 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         if self.src_ip is not None:
             result['SrcIP'] = self.src_ip
 
+        if self.terminal_info_collect_time is not None:
+            result['TerminalInfoCollectTime'] = self.terminal_info_collect_time
+
         if self.update_time is not None:
             result['UpdateTime'] = self.update_time
 
@@ -261,8 +322,23 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         if m.get('AutoLoginStatus') is not None:
             self.auto_login_status = m.get('AutoLoginStatus')
 
+        if m.get('BatteryHealthPercentage') is not None:
+            self.battery_health_percentage = m.get('BatteryHealthPercentage')
+
+        if m.get('BatteryRemainingPercentage') is not None:
+            self.battery_remaining_percentage = m.get('BatteryRemainingPercentage')
+
         if m.get('CPU') is not None:
             self.cpu = m.get('CPU')
+
+        if m.get('City') is not None:
+            self.city = m.get('City')
+
+        if m.get('Continent') is not None:
+            self.continent = m.get('Continent')
+
+        if m.get('Country') is not None:
+            self.country = m.get('Country')
 
         if m.get('CreateTime') is not None:
             self.create_time = m.get('CreateTime')
@@ -291,6 +367,12 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         if m.get('Disk') is not None:
             self.disk = m.get('Disk')
 
+        if m.get('DiskAvailable') is not None:
+            self.disk_available = m.get('DiskAvailable')
+
+        if m.get('DiskUsed') is not None:
+            self.disk_used = m.get('DiskUsed')
+
         if m.get('DlpStatus') is not None:
             self.dlp_status = m.get('DlpStatus')
 
@@ -312,6 +394,9 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         if m.get('InnerIP') is not None:
             self.inner_ip = m.get('InnerIP')
 
+        if m.get('JoinAdDomain') is not None:
+            self.join_ad_domain = m.get('JoinAdDomain')
+
         if m.get('Mac') is not None:
             self.mac = m.get('Mac')
 
@@ -332,6 +417,15 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
 
         if m.get('PaStatus') is not None:
             self.pa_status = m.get('PaStatus')
+
+        self.processes = []
+        if m.get('Processes') is not None:
+            for k1 in m.get('Processes'):
+                temp_model = main_models.GetUserDeviceResponseBodyDeviceProcesses()
+                self.processes.append(temp_model.from_map(k1))
+
+        if m.get('Province') is not None:
+            self.province = m.get('Province')
 
         if m.get('SaseUserId') is not None:
             self.sase_user_id = m.get('SaseUserId')
@@ -357,6 +451,9 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
         if m.get('SrcIP') is not None:
             self.src_ip = m.get('SrcIP')
 
+        if m.get('TerminalInfoCollectTime') is not None:
+            self.terminal_info_collect_time = m.get('TerminalInfoCollectTime')
+
         if m.get('UpdateTime') is not None:
             self.update_time = m.get('UpdateTime')
 
@@ -365,6 +462,57 @@ class GetUserDeviceResponseBodyDevice(DaraModel):
 
         if m.get('Workshop') is not None:
             self.workshop = m.get('Workshop')
+
+        return self
+
+class GetUserDeviceResponseBodyDeviceProcesses(DaraModel):
+    def __init__(
+        self,
+        cpu: float = None,
+        description: str = None,
+        memory: int = None,
+        name: str = None,
+    ):
+        self.cpu = cpu
+        self.description = description
+        self.memory = memory
+        self.name = name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.cpu is not None:
+            result['Cpu'] = self.cpu
+
+        if self.description is not None:
+            result['Description'] = self.description
+
+        if self.memory is not None:
+            result['Memory'] = self.memory
+
+        if self.name is not None:
+            result['Name'] = self.name
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Cpu') is not None:
+            self.cpu = m.get('Cpu')
+
+        if m.get('Description') is not None:
+            self.description = m.get('Description')
+
+        if m.get('Memory') is not None:
+            self.memory = m.get('Memory')
+
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
 
         return self